pyproject.toml
[tool.poetry]
name = "coreserver-ip-setter"
version = "1.0.0"
description = ""
authors = ["C. Plug <clpsplug@clpsplug.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
flake8 = "^7.0.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
